Port new build system to Tau Leaping C++ solver
- Renamed source files to match Makefile targets
- Add dependency to template file
- Port variable support to Tau Leaping and SSA solvers
